Enhance various layers and operations with improved error handling, dtype management, and performance optimizations
- Implement error handling for unsupported parameters in MultiheadAttention.
- Add tracking for the number of batches in BatchNorm layers and ensure proper updates.
- Improve dtype-aware handling in Embedding initialization and weight zeroing.
- Update Linear layer weight initialization to match PyTorch's Kaiming uniform.
- Optimize normalization operations with a numerically stable two-pass algorithm.
- Refactor pooling layers to store indices as Int64 for precision.
- Simplify RNNCell forward computation by removing unnecessary Variable wrapping.
- Enhance Module's to() method to only convert floating-point tensors.
- Modify arange function to use double for start, end, and step parameters, improving precision.
- Ensure contiguous tensors are used in matmul and related operations for performance.
- Validate broadcast shapes in logical operations to prevent runtime errors.

+++ include/tenzor/ops/creation.hpp
@@ -171,7 +171,7 @@ auto randint(int64_t low, int64_t high, std::vector<int64_t> shape,
  * auto t = arange(0.0f, 10.0f, 2.0f);  // [0, 2, 4, 6, 8]
  * @endcode
  */
-auto arange(float start, float end, float step = 1.0f,
+auto arange(double start, double end, double step = 1.0,
            DType dtype = DType::Float32,
            Device device = Device::cpu()) -> Tensor;
